feat: Add navigation component and bottom tab navigation for main app routes
