+++ app/src/main/java/com/uniandes/vinilosapp/views/MainActivity.kt
@@ -3,15 +3,11 @@ package com.uniandes.vinilosapp.views
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import com.uniandes.vinilosapp.views.album.Navigation
-import com.uniandes.vinilosapp.views.album.AlbumsScreen
-//import com.uniandes.vinilosapp.views.ui.theme.VinilosAppTheme
+import com.uniandes.vinilosapp.views.navigation.Navigation
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContent {
-            Navigation()
-        }
+        setContent { Navigation() }
     }
 }
